requirements.txt
# filepath: /workspaces/Trabalho-Breno/requirements.txt
grpcio
grpcio-tools
protobuf
pytest
